firebase/python/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

firebase-admin==6.2.0
paho-mqtt==1.6.1
requests==2.31.0
